refactor: move read-only profile queries from server actions to data layer
